// src/app/model/class-lesson-dto.ts
export interface ClassLessonDto {
    'id': number;
    'display_name': string;
    'class_lesson_id': number;
    'theme_lesson_id': number;
    'date': string;
    'start_datetime': string;
    'end_datetime': string;
    'theme_id': number;
    'theme_name': string;
    'enrolment_id': number;
    'has_attended': number;
    'remarks': string;
    'status': string;
    'created_at': string;
    'updated_at': string;
}